enc.h: designed the new API!
Using enc->dictID as a hash seed is tricky.  Why could this be needed?
My thinking behind it is that, well, what if the hash function turns out
to be poor, or we use too few bits, and this renders, with very small
probability, the whole thing so broken that everybody laughs at it?
Then we can simply rebuild the dictionary, and the problem goes away.
It is not an elegant solution, but not a catastrophe either, given the
alternative of changing the format.  In other words, rather than using
a hardcoded constant, it is better to employ as a seed something that
can be munged.